requirements.txt
biopython>=1.81
pandas>=2.0.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
jupyter>=1.0.0